% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_resource_to_dataset.R
\name{attach_resource_to_dataset}
\alias{attach_resource_to_dataset}
\title{attach_resource_to_dataset}
\usage{
attach_resource_to_dataset(resource_nid, file_path, credentials = list(cookie
  = dkanr::get_cookie(), token = dkanr::get_token()), attach = 1)
}
\arguments{
\item{resource_nid}{character: Data Catalog Resource Node ID}

\item{file_path}{character: Path to file on local machine}

\item{credentials}{Optional list parameter. Default values are Cookie and Token generated by dkan_setup()}

\item{attach}{numeric: Setting the attach parameter to 0 ensures that the file will replace any existing file on the resource. Setting it to 1 will result in a rejected request if the resource already has an attached file (but it will work if the resource file upload field is empty).}
}
\description{
Attach a file to an existing resource
}
