% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-attach_resources_template.R
\docType{data}
\name{attach_resources_template}
\alias{attach_resources_template}
\title{template to attach resource(s) to a dataset}
\format{A list of length 1.}
\usage{
attach_resources_template
}
\description{
Small template to generate the correct JSON file to attach resource(s) to an existing dataset
}
\examples{
attach_resources_template
}
\keyword{datasets}
