% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_resource_to_dataset.R
\name{attach_resources_to_dataset}
\alias{attach_resources_to_dataset}
\title{attach_resources_to_dataset}
\usage{
attach_resources_to_dataset(dataset_nid, resource_nids,
  root_url = dkanr::get_url(), credentials = list(cookie =
  dkanr::get_cookie(), token = dkanr::get_token()))
}
\arguments{
\item{dataset_nid}{character: node id of the dataset to which the resources are to be attached}

\item{resource_nids}{character vector: node ids of the resources to be attached}

\item{root_url}{character: API root URL}

\item{credentials}{list: authentication token and cookie}
}
\value{
list
}
\description{
Attach resources to a dataset in DDH
}
