% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_blank_json_body.R
\name{create_blank_json_body}
\alias{create_blank_json_body}
\title{create_blank_json_body}
\usage{
create_blank_json_body(
  values = c("title", "body"),
  publication_status = "published",
  type = "dataset",
  root_url = dkanr::get_url()
)
}
\arguments{
\item{values}{vector: vector of fields to make blank}

\item{publication_status}{string: status to post the mode, takes values of c("published", "unpublished", "draft")}

\item{type}{string: indicates whether the JSON body is for a "dataset" or a "resource"}

\item{root_url}{string: API root URL}
}
\value{
json string
}
\description{
Create the JSON body for updating the given fields or creating a new dataset
}
