% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{create_dataset}
\usage{
create_dataset(credentials, body, root_url = production_root_url)
}
\arguments{
\item{credentials}{list: object returned by the get_credentials() function}

\item{body}{json: json object generated by jsonlite::toJSON}

\item{root_url}{character: API root URL}
}
\value{
list
}
\description{
Create new dataset in DDH
}
