% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_json_attach.R
\name{create_json_attach}
\alias{create_json_attach}
\title{create_json_attach}
\usage{
create_json_attach(resource_nids, root_url = dkanr::get_url())
}
\arguments{
\item{resource_nids}{character: node ids of the resources}

\item{root_url}{string: API root URL}
}
\value{
json
}
\description{
Create a json according to predefined template
}
