% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_json_body.R
\name{create_json_body}
\alias{create_json_body}
\title{create_json_body}
\usage{
create_json_body(values = list(title = "Test Create JSON", body =
  "Test Creation of JSON"),
  json_formats = ddhconnect::dataset_json_format_lookup,
  lovs = ddhconnect::get_lovs(), root_url = dkanr::get_url())
}
\arguments{
\item{values}{list: list of corresponding values that need to be updated}

\item{json_formats}{dataframe: lookup table for machine name and expected value (dataset or resource)}

\item{root_url}{string: API root URL}
}
\value{
json string
}
\description{
Create the JSON body for updating the given fields or creating a new dataset
}
