% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_json_dataset.R
\name{create_json_dataset}
\alias{create_json_dataset}
\title{create_json_dataset}
\usage{
create_json_dataset(values = list(title = "Test Create JSON", body =
  "Test Creation of JSON", field_wbddh_dsttl_upi = "123", field_topic =
  "Poverty", field_wbddh_country = c("Antigua and Barbuda", "Armenia")),
  publication_status = "published",
  ddh_fields = ddhconnect::get_fields(), lovs = ddhconnect::get_lovs(),
  root_url = dkanr::get_url())
}
\arguments{
\item{values}{list: list of corresponding values that need to be updated}

\item{publication_status}{string: status to post the mode, takes values of c("published", "unpublished", "draft")}

\item{ddh_fields}{dataframe: table of all the data catalog fields by node type}

\item{lovs}{dataframe: lookup table of the data catalog tids and values}

\item{root_url}{string: API root URL}
}
\value{
json string
}
\description{
Create the JSON body for updating the given fields or creating a new dataset
}
