% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_json_resource.R
\name{create_json_resource}
\alias{create_json_resource}
\title{create_json_resource}
\usage{
create_json_resource(values = list(title = "Test Resource Title", body =
  "Test Resource Body", field_wbddh_data_class = "Public",
  field_wbddh_resource_type = "Resource Type not specified", field_format =
  "Format Not Specified", field_link_api = "www.google.com",
  field_ddh_harvest_src = "Finances", field_ddh_harvest_sys_id =
  "8675309"), workflow_status = "published",
  ddh_fields = ddhconnect::get_fields(), lovs = ddhconnect::get_lovs(),
  root_url = dkanr::get_url())
}
\arguments{
\item{values}{list: list of corresponding values that need to be updated}

\item{workflow_status}{string: status to post the mode, takes values of c("published", "unpublished", "draft")}

\item{ddh_fields}{dataframe: table of all the data catalog fields by node type}

\item{lovs}{dataframe: lookup table of the data catalog tids and values}

\item{root_url}{string: API root URL}
}
\value{
json string
}
\description{
Create the JSON body for updating the given fields or creating a new dataset
}
