% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_resource.R
\name{create_resource}
\alias{create_resource}
\title{create_resource}
\usage{
create_resource(credentials, body, root_url = production_root_url)
}
\arguments{
\item{credentials}{list: object returned by the get_credentials() function}

\item{body}{json: json object generated by jsonlite::toJSON}

\item{root_url}{character: API root URL}
}
\value{
list
}
\description{
Create new resource in DDH
}
