% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset_json_format_lookup}
\alias{dataset_json_format_lookup}
\title{Lookup Table of JSON Formats for Datasets}
\format{
A data frame with 73 rows and 2 variables
}
\usage{
data(dataset_json_format_lookup)
}
\description{
A dataset with JSON formatting information for datasets when
 creating or updating. The variables are:
}
\details{
\itemize{
 \item machine_names. Data Catalog field names
 \item json_template. JSON Format
}
}
\keyword{datasets}
