% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_dataset.R
\name{delete_dataset}
\alias{delete_dataset}
\title{delete_dataset}
\usage{
delete_dataset(nid, root_url = dkanr::get_url(),
  credentials = list(cookie = dkanr::get_cookie(), token =
  dkanr::get_token()))
}
\arguments{
\item{nid}{character: Node ID of the dataset to be updated}

\item{root_url}{character: API root URL}

\item{credentials}{list: authentication token and cookie}
}
\value{
list
}
\description{
Delete existing dataset in DDH
}
