% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_resource.R
\name{download_resource}
\alias{download_resource}
\title{download_resource}
\usage{
download_resource(resource_id, credentials = list(cookie =
  dkanr::get_cookie(), token = dkanr::get_token()))
}
\arguments{
\item{resource_id}{character: Resource NID}

\item{credentials}{Optional list parameter. Default values are Cookie and Token generated by dkan_setup()}
}
\description{
Download a file, extracting the path from `get_metadata()` output
}
