% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_credentials.R
\name{get_credentials}
\alias{get_credentials}
\title{get_credentials}
\usage{
get_credentials(username, password)
}
\arguments{
\item{username}{character: Your DDH API username}

\item{password}{character: Your DDH API password}
}
\value{
list
}
\description{
This function facilitate the authentication process with the DDH API.
}
