% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets_count.R
\name{get_datasets_count}
\alias{get_datasets_count}
\title{get_datasets_count()}
\usage{
get_datasets_count(
  datatype = "All",
  root_url = dkanr::get_url(),
  credentials = list(cookie = dkanr::get_cookie(), token = dkanr::get_token())
)
}
\arguments{
\item{datatype}{character: Restrict the list to a specific type of dataset. Available options are: "All", "Timeseries", "Microdata", "Geospatial", or "Other"}

\item{root_url}{character: API root URL}

\item{credentials}{list: API authentication credentials}
}
\value{
numeric vector
}
\description{
get_datasets_count()
}
