% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets_list.R
\name{get_datasets_list}
\alias{get_datasets_list}
\title{get_datasets_list()}
\usage{
get_datasets_list(datatype = c("all", "timeseries", "microdata", "geospatial",
  "other"))
}
\arguments{
\item{datatype}{character: Restrict the list to a specific type of dataset. Available options are: "all", "timeseries", "geospatial", or "other"}
}
\value{
numeric vector
}
\description{
get_datasets_list()
}
