% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fields.R
\name{get_fields}
\alias{get_fields}
\title{get_fields}
\usage{
get_fields(root_url = dkanr::get_url(), credentials = list(cookie =
  dkanr::get_cookie(), token = dkanr::get_token()))
}
\arguments{
\item{root_url}{character: API root URL}

\item{credentials}{list: authentication token and cookie}
}
\value{
dataframe
}
\description{
Retrieve list of fields (machine names + screen names) for each data type
}
