% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iso3.R
\name{get_iso3}
\alias{get_iso3}
\title{get_iso3}
\usage{
get_iso3(root_url = dkanr::get_url())
}
\arguments{
\item{root_url}{character: API root URL}
}
\value{
dataframe
}
\description{
Retrieve list of country names and iso3 codes accepted by DDH
}
