% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iso3.R
\name{get_iso3_codes}
\alias{get_iso3_codes}
\title{get_iso3_codes}
\usage{
get_iso3_codes(root_url = dkanr::get_url())
}
\arguments{
\item{root_url}{character: API root URL}
}
\value{
dataframe
}
\description{
Non memoised version: Retrieve list of country names and iso3 codes
accepted by DDH
}
