% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lov_fields.R
\name{get_lov_fields}
\alias{get_lov_fields}
\title{get_lov_fields}
\usage{
get_lov_fields(root_url = dkanr::get_url())
}
\arguments{
\item{root_url}{string: API root URL}
}
\value{
vector
}
\description{
Retrieve list of lov fields
}
