% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lovs.R
\name{get_lovs}
\alias{get_lovs}
\title{get_lovs}
\usage{
get_lovs(root_url = production_root_url)
}
\arguments{
\item{root_url}{character: API root URL}
}
\value{
dataframe
}
\description{
Retrieve controlled vocabulary / list of values (LOVs) accepted by DDH
}
