% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_required_fields.R
\name{get_required_fields}
\alias{get_required_fields}
\title{get_required_fields}
\usage{
get_required_fields(datatype = "Other")
}
\arguments{
\item{datatype}{character: One value out of "Timeseries", "Microdata", "Geospatial" and "Other"}
}
\value{
vector
}
\description{
Retrieve list of required fields for a data type
}
