% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resource_nid.R
\name{get_resource_nid}
\alias{get_resource_nid}
\title{get_resource_nid}
\usage{
get_resource_nid(nid, credentials, root_url = production_root_url)
}
\arguments{
\item{nid}{character: The DATASET node id}

\item{credentials}{list: object returned by the get_credentials() function}

\item{root_url}{character: API root URL}
}
\value{
character
}
\description{
Return ALL resources node IDs associated with a dataset
}
