% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_details.R
\name{get_user_details}
\alias{get_user_details}
\title{get_user_details}
\usage{
get_user_details(credentials = list(cookie = dkanr::get_cookie(), token =
  dkanr::get_token()), uid, root_url = dkanr::get_url())
}
\arguments{
\item{credentials}{list: authentication token and cookie}

\item{uid}{numeric: UPI number}

\item{root_url}{character: API root URL}
}
\value{
character
}
\description{
Retrieve user details: first name, last name, DDH internal ID
}
