% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout_ddh.R
\name{logout_ddh}
\alias{logout_ddh}
\title{logout_ddh}
\usage{
logout_ddh(credentials, root_url = production_root_url)
}
\arguments{
\item{credentials}{list: object returned by the get_credentials() function}

\item{root_url}{character: Server's root URL}
}
\value{
list
}
\description{
Close an existing connection with the DDH API
}
