% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_metadata_excel.R
\name{map_metadata_excel}
\alias{map_metadata_excel}
\title{map_metadata_excel}
\usage{
map_metadata_excel(path)
}
\arguments{
\item{path}{string: Takes path to an excel file containing metadata}
}
\value{
vector
}
\description{
Extract metadata from an excel file. Map the values to the corresponding machine_names.
The file has to have "Metadata field" and "Values" columns.
}
