% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tids.R
\name{map_tids}
\alias{map_tids}
\title{map_tids}
\usage{
map_tids(values, credentials = list(cookie = dkanr::get_cookie(), token =
  dkanr::get_token()))
}
\arguments{
\item{values}{character vector: list of corresponding values that need to be checked for tids}

\item{credentials}{(optional) list: authentication token and cookie}
}
\value{
values character vector
}
\description{
Map strings to tids
}
