% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dkanr_setup}
\alias{dkanr_settings}
\alias{dkanr_logout}
\alias{get_resource_nids}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dkanr}{\code{\link[dkanr]{dkanr_setup}}, \code{\link[dkanr]{dkanr_settings}}, \code{\link[dkanr]{dkanr_logout}}, \code{\link[dkanr]{get_resource_nids}}}
}}

