% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_catalog.R
\name{search_catalog}
\alias{search_catalog}
\title{search_catalog}
\usage{
search_catalog(fields = c("nid", "uuid", "title", "field_contact_email",
  "field_wbddh_data_type", "status"), filters = c(field_wbddh_data_type = 294,
  status = 1), limit = 200, credentials,
  path = "search-service/search_api/datasets",
  root_url = production_root_url)
}
\arguments{
\item{fields}{character vector: fields to be returned by the search}

\item{filters}{named character vector: filters to be applied to the search}

\item{limit}{numeric: Maximum number of records to retrieve by search iteration}

\item{credentials}{list: API authentication credentials}

\item{path}{character: path to the search_api}

\item{root_url}{character: API root URL}
}
\value{
list
}
\description{
Helper function to search the DDH data catalog
}
