% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_dataset_update_json}
\alias{test_dataset_update_json}
\title{Test Values for Dataset Update}
\format{A list with 30 items}
\usage{
data(test_dataset_update_json)
}
\description{
A list with nested content for a test dataset's update
}
\keyword{datasets}
