% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_resource_update_json}
\alias{test_resource_update_json}
\title{Test Values for Resource Update}
\format{A list with 6 items}
\usage{
data(test_resource_update_json)
}
\description{
A list with nested content for a test resource's update
}
\keyword{datasets}
