% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ui_names_lookup}
\alias{ui_names_lookup}
\title{Lookup Table for User Interface Form Fields and JSON Field Names}
\format{A data frame with 115 rows and 2 variables}
\usage{
data(ui_names_lookup)
}
\description{
A dataset with UI form names and corresponding field names
}
\details{
\itemize{
 \item machine_names. Data Catalog field names
 \item ui_names. form name
}
}
\keyword{datasets}
