% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_resource.R
\name{update_resource}
\alias{update_resource}
\title{update_resource}
\usage{
update_resource(credentials, nid, body, root_url = production_root_url)
}
\arguments{
\item{credentials}{list: object returned by the get_credentials() function}

\item{nid}{character: Node ID of the resource to be updated}

\item{body}{json: json object generated by jsonlite::toJSON}

\item{root_url}{character: API root URL}
}
\value{
list
}
\description{
Update existing resource in DDH
}
