% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attach_resources_template}
\alias{attach_resources_template}
\title{Format to Attach a Resource}
\format{A list with 1 item}
\usage{
data(attach_resources_template)
}
\description{
A list with format needed to attach a resource to a dataset
}
\keyword{datasets}
