% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{create_dataset}
\usage{
create_dataset(credentials, path)
}
\arguments{
\item{credentials}{list: object returned by the get_credentials() function}

\item{path}{character: Path to .JSON file containing the dataset information to be posted on DDH}
}
\value{
list
}
\description{
Create new dataset in DDH
}
