% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{create_dataset}
\usage{
create_dataset(root_url = dkanr::get_url(), body, credentials = list(cookie
  = dkanr::get_cookie(), token = dkanr::get_token()))
}
\arguments{
\item{root_url}{character: API root URL}

\item{body}{json: JSON object generated by jsonlite::toJSON}

\item{credentials}{list: authentication token and cookie}
}
\value{
list
}
\description{
Create new dataset in DDH
}
