% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{create_dataset}
\usage{
create_dataset(
  body,
  root_url = dkanr::get_url(),
  credentials = list(cookie = dkanr::get_cookie(), token = dkanr::get_token())
)
}
\arguments{
\item{body}{json: JSON object generated by jsonlite::toJSON}

\item{root_url}{character: API root URL}

\item{credentials}{list: authentication token and cookie}
}
\value{
list
}
\description{
Create new dataset in DDH
}
