% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_json_attach.R
\name{create_json_attach}
\alias{create_json_attach}
\title{create_json_attach}
\usage{
create_json_attach(resource_nids)
}
\arguments{
\item{resource_nid}{character: resource node id}

\item{json_template}{list: List generated from JSON template}
}
\value{
json
}
\description{
Create a json according to predefined template
}
