% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_json_body.R
\name{create_json_body}
\alias{create_json_body}
\title{create_json_body}
\usage{
create_json_body(values = list(title = "Test Create JSON", body =
  "Test Creation of JSON", field_wbddh_dsttl_upi = "123", field_topic =
  "Poverty", field_wbddh_country = c("Antigua and Barbuda", "Armenia")),
  node_type = "dataset", root_url = dkanr::get_url())
}
\arguments{
\item{values}{list: list of corresponding values that need to be updated}

\item{node_type}{string: type of node to upload (dataset or resource)}

\item{root_url}{string: API root URL}
}
\value{
json string
}
\description{
Create the JSON body for updating the given fields or creating a new dataset
}
