% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_resource.R
\name{create_resource}
\alias{create_resource}
\title{create_resource}
\usage{
create_resource(
  body,
  root_url = dkanr::get_url(),
  credentials = list(cookie = dkanr::get_cookie(), token = dkanr::get_token())
)
}
\arguments{
\item{body}{json: json object generated by jsonlite::toJSON}

\item{root_url}{character: API root URL}

\item{credentials}{list: authentication token and cookie}
}
\value{
list
}
\description{
Create new resource in DDH
}
