% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_resource.R
\name{create_resource}
\alias{create_resource}
\title{create_resource}
\usage{
create_resource(credentials, path)
}
\arguments{
\item{credentials}{list: object returned by the get_credentials() function}

\item{path}{character: Path to .JSON file containing the resource information to be posted on DDH}
}
\value{
list
}
\description{
Create new resource in DDH
}
