% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_resource.R
\name{download_resource}
\alias{download_resource}
\title{download_resource}
\usage{
download_resource(resource_nid, destination = getwd(),
  root_url = dkanr::get_url(), credentials = list(cookie =
  dkanr::get_cookie(), token = dkanr::get_token()))
}
\arguments{
\item{resource_nid}{character: Resource NID}

\item{destination}{string: file path to save the file, defaults to current working directory}

\item{root_url}{string: API root URL}

\item{credentials}{list: authentication token and cookie}
}
\description{
Download a resource file, extracting the path from `get_metadata()` output
}
