% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_resource.R
\name{download_resource}
\alias{download_resource}
\title{download_resource}
\usage{
download_resource(resource_nid, credentials = list(cookie =
  dkanr::get_cookie(), token = dkanr::get_token()))
}
\arguments{
\item{resource_nid}{character: Resource NID}

\item{credentials}{list: authentication token and cookie}
}
\description{
Download a resource file, extracting the path from `get_metadata()` output
}
