% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lovs.R
\name{get_lovs}
\alias{get_lovs}
\title{get_lovs}
\usage{
get_lovs(credentials)
}
\arguments{
\item{credentials}{}
}
\value{
dataframe
}
\description{
Retrieve controlled vocabulary / list of values (LOVs) accepted by DDH
}
