% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{get_metadata}
\usage{
get_metadata(id)
}
\arguments{
\item{id}{character: The dataset UUID}
}
\value{
list
}
\description{
Retrieve metadata for a specific dataset
}
