% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{get_metadata}
\usage{
get_metadata(credentials, node)
}
\arguments{
\item{credentials}{list: object returned by get_credentials()}

\item{node}{numeric: The dataset node}
}
\value{
dataframe
}
\description{
Retrieve metadata for a specific dataset
}
