% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{get_metadata}
\usage{
get_metadata(nid, credentials = list(cookie = dkanr::get_cookie(), token =
  dkanr::get_token()), root_url = dkanr::get_url())
}
\arguments{
\item{nid}{character: The dataset node id}

\item{credentials}{list: authentication token and cookie}

\item{root_url}{character: API root URL}
}
\value{
list
}
\description{
Retrieve metadata for a specific dataset or resource
}
