% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{get_metadata}
\usage{
get_metadata(id, root_url = production_root_url)
}
\arguments{
\item{id}{character: The dataset UUID}

\item{root_url}{character: API root URL}
}
\value{
list
}
\description{
Retrieve metadata for a specific dataset
}
