% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resource_nid.R
\name{get_resource_nid}
\alias{get_resource_nid}
\title{get_resource_nid}
\usage{
get_resource_nid(nid, root_url = dkanr::get_url(), credentials = list(cookie
  = dkanr::get_cookie(), token = dkanr::get_token()))
}
\arguments{
\item{nid}{character: The DATASET node id}

\item{root_url}{character: API root URL}
}
\value{
character
}
\description{
Return ALL resources node IDs associated with a dataset
}
