% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_details.R
\name{get_user_details}
\alias{get_user_details}
\title{get_user_details}
\usage{
get_user_details(credentials, uid, root_url = production_root_url)
}
\arguments{
\item{credentials}{list: object returned by the get_credentials() function}

\item{uid}{numeric: UPI number}

\item{root_url}{character: API root URL}
}
\value{
character
}
\description{
Retrieve user details: first name, last name, DDH internal ID
}
