% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_details.R
\name{get_user_details}
\alias{get_user_details}
\title{get_user_details}
\usage{
get_user_details(
  uid,
  root_url = dkanr::get_url(),
  credentials = list(cookie = dkanr::get_cookie(), token = dkanr::get_token())
)
}
\arguments{
\item{uid}{numeric: UPI number}

\item{root_url}{character: API root URL}

\item{credentials}{list: authentication token and cookie}
}
\value{
character
}
\description{
Retrieve user details: first name, last name, DDH internal ID
}
