% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout_ddh.R
\name{logout_ddh}
\alias{logout_ddh}
\title{logout_ddh}
\usage{
logout_ddh(root_url = dkanr::get_url(), credentials = list(cookie =
  dkanr::get_cookie(), token = dkanr::get_token()))
}
\arguments{
\item{root_url}{character: API root URL}

\item{credentials}{list: authentication token and cookie}
}
\value{
list
}
\description{
Close an existing connection with the DDH API
}
