% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout_ddh.R
\name{logout_ddh}
\alias{logout_ddh}
\title{logout_ddh}
\usage{
logout_ddh(credentials = list(cookie = dkanr::get_cookie(), token =
  dkanr::get_token()), root_url = dkanr::get_url())
}
\arguments{
\item{credentials}{list: authentication token and cookie}

\item{root_url}{character: Server's root URL}
}
\value{
list
}
\description{
Close an existing connection with the DDH API
}
