% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tids.R
\name{map_tids}
\alias{map_tids}
\title{map_tids}
\usage{
map_tids(metadata)
}
\arguments{
\item{resource_id}{character: Resource NID}

\item{credentials}{Optional list parameter. Default values are Cookie and Token generated by dkan_setup()}
}
\description{
Map strings to tids
}
