% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tids.R
\name{map_tids}
\alias{map_tids}
\title{map_tids}
\usage{
map_tids(values, root_url = dkanr::get_url())
}
\arguments{
\item{values}{character vector: list of corresponding values that need to be checked for tids}

\item{root_url}{character: API root URL}
}
\value{
list: list of metadata with some of the fields to be converted to
}
\description{
Map strings to tids
}
