% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_catalog.R
\name{search_catalog}
\alias{search_catalog}
\title{search_catalog}
\usage{
search_catalog(fields = c("nid", "uuid", "title", "field_contact_email",
  "field_wbddh_data_type", "status"), filters = c(field_wbddh_data_type =
  294, status = 1), limit = 200, root_url = dkanr::get_url(),
  credentials = list(cookie = dkanr::get_cookie(), token =
  dkanr::get_token()))
}
\arguments{
\item{fields}{character vector: fields to be returned by the search}

\item{filters}{named character vector: filters to be applied to the search}

\item{limit}{numeric: Maximum number of records to retrieve by search iteration}

\item{root_url}{character: API root URL}

\item{credentials}{list: API authentication credentials}
}
\value{
list
}
\description{
Helper function to search the DDH data catalog
}
