% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ddh.R
\name{search_ddh}
\alias{search_ddh}
\title{search_ddh}
\usage{
search_ddh(credentials,
  query = "limit=20&fields=[,nid,uuid,title,]&filter[field_wbddh_data_type]=294",
  root_url = production_root_url)
}
\arguments{
\item{credentials}{list: object returned by the get_credentials() function}

\item{query}{character: full text query}

\item{root_url}{character: API root URL}
}
\value{
list
}
\description{
Search DHH using the search_api service
}
