% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dataset.R
\name{update_dataset}
\alias{update_dataset}
\title{update_dataset}
\usage{
update_dataset(credentials, nid, body, root_url = production_root_url)
}
\arguments{
\item{credentials}{list: object returned by the get_credentials() function}

\item{nid}{character: Node ID of the dataset to be updated}

\item{body}{json: json object generated by jsonlite::toJSON}

\item{root_url}{character: API root URL}
}
\value{
list
}
\description{
Update existing in DDH
}
