% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dataset.R
\name{update_dataset}
\alias{update_dataset}
\title{update_dataset}
\usage{
update_dataset(
  nid,
  body,
  root_url = dkanr::get_url(),
  credentials = list(cookie = dkanr::get_cookie(), token = dkanr::get_token())
)
}
\arguments{
\item{nid}{character: Node ID of the dataset to be updated}

\item{body}{json: json object generated by jsonlite::toJSON}

\item{root_url}{character: API root URL}

\item{credentials}{list: authentication token and cookie}
}
\value{
list
}
\description{
Update existing in DDH
}
