% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_resource.R
\name{update_resource}
\alias{update_resource}
\title{update_resource}
\usage{
update_resource(
  credentials = list(cookie = dkanr::get_cookie(), token = dkanr::get_token()),
  nid,
  body,
  root_url = dkanr::get_url()
)
}
\arguments{
\item{credentials}{list: authentication token and cookie}

\item{nid}{character: Node ID of the resource to be updated}

\item{body}{json: json object generated by jsonlite::toJSON}

\item{root_url}{character: API root URL}
}
\value{
list
}
\description{
Update existing resource in DDH
}
